integrated UserDetailsServiceImpl one method into UserService
